il est temps de se tirer une balle

+++ exercices/Cursus/printf/bonus/utils2.c
@@ -6,13 +6,13 @@
 /*   By: locagnio <locagnio@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 19:53:23 by locagnio          #+#    #+#             */
-/*   Updated: 2024/11/14 16:18:07 by locagnio         ###   ########.fr       */
+/*   Updated: 2024/11/15 16:43:09 by locagnio         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_digits(int n)
+int	ft_digits(long long n)
 {
 	int count;
